FP.5 Completed
The method receive is correctly implemented

The method receive should use std::unique_lock<std::mutex> and
_condition.wait() to wait for and receive new messages and pull them from the queue using move semantics. The received object should then be returned by the receive function.

The method waitForGreen is correctly implemented

The method waitForGreen is completed, in which an infinite while loop runs and repeatedly calls the receive function on the message queue. Once it receives TrafficLightPhase::green, the method returns.